增加后续处理浮窗
